+++ aa_modules/aamod_waveletdespike.m
@@ -48,9 +48,16 @@ switch task
         P = spm_file(P,'prefix','m');
         
         %% Despike
+        if isfield(aap.tasklist.currenttask.settings,'chainsearch')
+            chsearch = aap.tasklist.currenttask.settings.chainsearch;
+        else
+            chsearch = 'moderate';
+        end
         mem = meminfo;
         outfilepfx = horzcat(spm_file(P,'path'),repmat(filesep,[size(P,1) 1]),spm_file(P,'basename'));
-        WaveletDespike(P,outfilepfx, 'LimitRAM', min([MAXMEM mem.MemFree*0.9]));
+        WaveletDespike(P,outfilepfx, ...
+            'chsearch', chsearch,...
+            'LimitRAM', min([MAXMEM mem.MemFree*0.9]));
         outputfns = spm_file(outfilepfx,'suffix','_wds.nii');
         
         %% Describe outputs